+++ POGOLib.Official/Net/RpcEncryption.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Threading.Tasks;
 using Google.Protobuf;
 using POGOLib.Official.Extensions;
-using POGOLib.Official.Logging;
 using POGOLib.Official.Util;
 using POGOProtos.Networking.Envelopes;
 using POGOProtos.Networking.Platform;
@@ -17,7 +16,6 @@ namespace POGOLib.Official.Net
 {
     internal class RpcEncryption
     {
-
         /// <summary>
         /// The authenticated <see cref="Session"/>.
         /// </summary>
@@ -51,8 +49,7 @@ namespace POGOLib.Official.Net
         }
 
         private long TimestampSinceStartMs => _stopwatch.ElapsedMilliseconds;
-
-
+        
         /// <summary>
         /// Generates a few random <see cref="LocationFix"/>es to act like a real GPS sensor.
         /// </summary>
@@ -78,7 +75,7 @@ namespace POGOLib.Official.Net
             //      Not so relevant when starting up.
             var totalMilliseconds = providerCount * millisecondsPerFix;
             var baseTimestampSnapshot = Math.Max(timestampSinceStart - totalMilliseconds, 0);
-            var playAroundWindow = baseTimestampSnapshot - _lastTimestampSinceStart;
+            var playAroundWindow = Math.Max(0, baseTimestampSnapshot - _lastTimestampSinceStart);
 
             if (playAroundWindow == 0 && providerCount == 1 && millisecondsPerFix >= timestampSinceStart)
             {
@@ -95,7 +92,7 @@ namespace POGOLib.Official.Net
             var playAroundWindowPart = playAroundWindow != 0
                 ? playAroundWindow / providerCount
                 : 1;
-
+            
             for (var i = 0; i < providerCount; i++)
             {
                 var timestampSnapshot = baseTimestampSnapshot;
@@ -103,7 +100,7 @@ namespace POGOLib.Official.Net
                 timestampSnapshot += i * millisecondsPerFix;
                 // Apply an offset.
                 timestampSnapshot += _session.Random.Next(0, (int) ((i + 1) * playAroundWindowPart));
-
+                
                 locationFixes.Add(new LocationFix
                 {
                     TimestampSnapshot = (ulong) timestampSnapshot,
@@ -173,7 +170,7 @@ namespace POGOLib.Official.Net
                         Status = 3
                     }
                 },
-                DeviceInfo = _session.DeviceInfo,
+                DeviceInfo = _session.Device.DeviceInfo,
                 LocationFix = { locationFixes },
                 ActivityStatus = new ActivityStatus
                 {
@@ -209,6 +206,5 @@ namespace POGOLib.Official.Net
             
             return encryptedSignature;
         }
-
     }
 }